+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### Unreleased
 
+- test: refactored against test-fixtures 1.7.0
+
 ### [2.1.0] - 2026-05-20
 
 - refactor: extract `normalize_redis_ini` / `normalize_endpoint` as helpers